MIM-139 拆分 Composer 与状态托盘测试职责
