+++ ios/MimiRemote/Sources/Features/Conversation/ComposerView.swift
@@ -1702,184 +1702,6 @@ struct ComposerView: View {
         .accessibilityHint(L10n.text("ui.adjust_speed_and_output"))
     }
 
-    @ViewBuilder
-    var modelPickerControl: some View {
-        modelPickerControl(showsTitle: true, usesCompactTitle: false)
-    }
-
-    @ViewBuilder
-    func modelPickerControl(showsTitle: Bool, usesCompactTitle: Bool = false) -> some View {
-        let visibleTitle = usesCompactTitle ? compactModelPickerTitle : modelPickerTriggerTitle
-        Button {
-            showsModelGridPicker.toggle()
-        } label: {
-            composerToolbarControlLabel(
-                // iPhone 用短标签外显模型与推理档位；完整名称仍通过 accessibilityValue 提供。
-                title: showsTitle ? visibleTitle : nil,
-                systemImage: usesCompactTitle ? nil : "cpu",
-                trailingSystemImage: ConversationLayout.compactComposerShowsFastModeIndicator(
-                    usesCompactMetrics: usesCompactComposerMetrics,
-                    isFastModeSelected: isFastModeSelected
-                ) ? "bolt.fill" : nil,
-                titleMaxWidth: usesCompactTitle
-                    ? ConversationLayout.compactComposerModelTitleMaxWidth(availableWidth: availableWidth)
-                    : (usesCompactComposerMetrics ? 82 : 150),
-                usesCondensedTitle: usesCompactTitle,
-                accessibilityLabel: L10n.text("ui.switch_model_and_inference_strength")
-            )
-            .contentTransition(.opacity)
-        }
-        .buttonStyle(MimiPressButtonStyle(reduceMotion: reduceMotion))
-        .accessibilityLabel(L10n.text("ui.switch_model_and_inference_strength"))
-        .accessibilityValue(modelShortcutAccessibilityValue(for: modelPickerTriggerTitle))
-        .accessibilityHint(L10n.text("ui.select_the_model_to_use_in_the_next"))
-        .accessibilityIdentifier("composer.model")
-        .popover(isPresented: $showsModelGridPicker, arrowEdge: .bottom) {
-            ModelReasoningGridPicker(
-                options: modelOptionsForMenu,
-                layout: modelReasoningGridLayout,
-                selection: selectedModelGridSelection,
-                selectedModelID: composerState.turnOptions.model,
-                isRefreshing: sessionStore.isRefreshingAppServerModels,
-                isFastMode: isFastModeSelected,
-                onSelectModel: { option, effort in
-                    selectModel(option, effort: effort)
-                },
-                onSelectDefaultModel: { option, effort in
-                    selectDefaultModel(option, effort: effort)
-                },
-                onFastModeChange: { isEnabled in
-                    composerState.updateTurnOptions {
-                        $0.serviceTier = ModelReasoningGridCatalog.serviceTierForFastMode(isEnabled)
-                    }
-                },
-                onRefresh: {
-                    Task { await sessionStore.refreshAppServerModelOptions(force: true) }
-                }
-            )
-            .environmentObject(themeStore)
-            .presentationCompactAdaptation(horizontal: .sheet, vertical: .sheet)
-            // compact popover 转成 sheet 后，iOS 26 不会稳定采用 fitted 内容高度。
-            // 标准字号只保留与实际模型行数一致的 detent；辅助功能字号则使用
-            // large detent，并继续由 Picker 内部滚动，避免放大文字被裁切。
-            .presentationDetents(modelPickerPresentationDetents)
-            .presentationDragIndicator(.visible)
-            .presentationBackground(themeStore.tokens(for: colorScheme).surface)
-        }
-    }
-
-    var modelPickerPresentationDetents: Set<PresentationDetent> {
-        if dynamicTypeSize.isAccessibilitySize {
-            return [.large]
-        }
-        return [.height(modelReasoningGridLayout.standardContentHeight)]
-    }
-
-    var effectiveModelID: String? {
-        ModelReasoningGridCatalog.effectiveModelID(
-            selectedModelID: composerState.turnOptions.model,
-            options: modelOptionsForMenu
-        )
-    }
-
-    var modelReasoningGridLayout: ModelReasoningGridLayout {
-        ModelReasoningGridCatalog.layout(
-            runtimeProvider: selectedSessionRuntimeProviderForModelMenu,
-            options: modelOptionsForMenu
-        )
-    }
-
-    var isFastModeSelected: Bool {
-        modelReasoningGridLayout.showsFastMode && composerState.turnOptions.serviceTier == "priority"
-    }
-
-    func modelShortcutAccessibilityValue(for title: String) -> String {
-        isFastModeSelected ? "\(title) · \(L10n.text("ui.fast"))" : title
-    }
-
-    var selectedModelGridSelection: ModelReasoningGridSelection {
-        let layout = modelReasoningGridLayout
-        let selectedOption = effectiveModelID.flatMap { modelID in
-            modelOptionsForMenu.first { $0.model.caseInsensitiveCompare(modelID) == .orderedSame }
-        }
-        let option = selectedOption
-            ?? layout.model(matching: effectiveModelID)
-            ?? layout.models.first(where: \.isDefault)
-            ?? layout.models.first
-        let effort = ModelReasoningGridCatalog.normalizedVisibleEffort(
-            option: option,
-            current: composerState.turnOptions.reasoningEffort,
-            layout: layout
-        )
-        return ModelReasoningGridSelection(
-            modelID: option?.model ?? "gpt-5.6-sol",
-            effort: effort
-        )
-    }
-
-    var modelPickerTriggerTitle: String {
-        guard let selectedModel = effectiveModelID,
-              let selectedEffort = developerModeEnabled
-                  ? composerState.turnOptions.reasoningEffort
-                  : selectedModelGridSelection.effort,
-              let title = ModelReasoningGridCatalog.triggerTitle(
-                  for: selectedModel,
-                  effort: selectedEffort,
-                  layout: modelReasoningGridLayout
-              )
-        else {
-            return selectedModelSummaryTitle
-        }
-        return title
-    }
-
-    var compactModelPickerTitle: String {
-        guard let selectedModel = effectiveModelID,
-              let selectedEffort = developerModeEnabled
-                  ? composerState.turnOptions.reasoningEffort
-                  : selectedModelGridSelection.effort,
-              let title = ModelReasoningGridCatalog.compactTriggerTitle(
-                  for: selectedModel,
-                  effort: selectedEffort,
-                  layout: modelReasoningGridLayout
-              )
-        else {
-            return selectedModelSummaryTitle
-        }
-        return title
-    }
-
-    func selectModel(
-        _ option: CodexAppServerModelOption,
-        effort: CodexAppServerReasoningEffort?
-    ) {
-        composerState.updateTurnOptions { options in
-            ModelReasoningGridCatalog.applySelection(
-                option: option,
-                effort: effort,
-                preservesServerDefault: false,
-                fallbackRuntimeProvider: payloadRuntimeProviderForSelectedSessionLock(),
-                to: &options
-            )
-        }
-    }
-
-    func selectDefaultModel(
-        _ option: CodexAppServerModelOption,
-        effort: CodexAppServerReasoningEffort?
-    ) {
-        composerState.updateTurnOptions { options in
-            // Default Model 只借用服务端默认模型的强度菜单，协议层继续提交 model = nil。
-            ModelReasoningGridCatalog.applySelection(
-                option: option,
-                effort: effort,
-                preservesServerDefault: true,
-                fallbackRuntimeProvider: payloadRuntimeProviderForSelectedSessionLock(),
-                to: &options
-            )
-        }
-    }
-
     var outputOptionsMenu: some View {
         Menu {
             Section(L10n.text("ui.summary")) {
